# internal/api/docs/swagger.yaml
components:
  schemas:
...
    api.AdminError:
      properties:
        error:
          type: string
        services:
          description: |-
            Services lists the service ids that still reference a group when a
            delete is refused with 409.
          items:
            type: string
          type: array
          uniqueItems: false
      type: object
...
    api.AdminPacksResponse:
      properties:
        argocdAvailable:
          description: |-
            ArgoCDAvailable is false when Applications could not be listed; packs
            are then derived from NebariApp labels only and carry no `argo` block.
          type: boolean
        error:
          type: string
        packs:
          items:
            $ref: '#/components/schemas/packs.Pack'
          type: array
          uniqueItems: false
      type: object
...
    packs.ArgoApp:
      properties:
        autoSync:
          type: boolean
        chart:
          type: string
        healthStatus:
          type: string
        images:
          items:
            type: string
          type: array
          uniqueItems: false
        lastSyncAt:
          type: string
        lastSyncPhase:
          type: string
        name:
          type: string
        namespace:
          type: string
        path:
          type: string
        reconciledAt:
          type: string
        repoURL:
          type: string
        resourceCount:
          type: integer
        revision:
          type: string
        syncStatus:
          type: string
        targetRevision:
          type: string
        tier:
          type: string
      type: object
    packs.Pack:
      properties:
        appVersion:
          type: string
        argo:
          $ref: '#/components/schemas/packs.ArgoApp'
        chartName:
          type: string
        chartVersion:
          type: string
        name:
          type: string
        namespace:
          type: string
        services:
          items:
            $ref: '#/components/schemas/packs.PackService'
          type: array
          uniqueItems: false
        tier:
          type: string
      type: object
    packs.PackService:
      properties:
        displayName:
          type: string
        healthStatus:
          type: string
        id:
          type: string
        name:
          type: string
        namespace:
          type: string
        requiredGroups:
          items:
            type: string
          type: array
          uniqueItems: false
        url:
          type: string
        visibility:
          type: string
      type: object
  securitySchemes:
    BearerAuth:
      description: Keycloak-issued JWT, sent as "Bearer <token>".
      in: header
      name: Authorization
      type: apiKey
...
info:
  description: |-
    Service-discovery and notifications API consumed by the nebari-landing SPA.
    Spec is published only when the server is started with --enable-docs (off in production).
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: Nebari Landing Page API
  version: v1
openapi: 3.1.0
paths:
...
  /admin/packs:
    get:
      description: One row per ArgoCD Application labeled part-of=nebari-packs (tier
        "pack") or nebari-foundational (tier "platform"), joined with the landing-page
        services it owns via the ArgoCD tracking-id annotation or destination namespace.
        Degrades to NebariApp Helm labels when ArgoCD is not readable.
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.AdminPacksResponse'
          description: OK
        "403":
          content:
            application/json:
              schema:
                type: string
          description: 'Forbidden: admin group required'
      security:
      - BearerAuth: []
      summary: List software packs (admin)
      tags:
      - admin
  /admin/packs/{name}:
    get:
      parameters:
      - description: Pack (ArgoCD Application) name
        in: path
        name: name
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/packs.Pack'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.AdminError'
          description: Not Found
      security:
      - BearerAuth: []
      summary: Get a software pack (admin)
      tags:
      - admin
...
servers:
- url: /api/v1